Update ugly imports + refactoring
