--- src/common/types.ts
@@ -1,14 +1,0 @@
-﻿export const enum PlaceType {
-  Apartment = 'Apartment',
-  Room = 'Room'
-}
-
-export type PlaceOffer = {
-  type: PlaceType;
-  isFavorite: boolean;
-  isPremium: boolean;
-  price: number;
-  rating: number;
-  title: string;
-  previewImage: string;
-};
