+++ src/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import App from './components/app/app.tsx';
-import {placeOffers} from './constants/data.ts';
+import App from '@/components/app/app.tsx';
+import {placeOffers} from '@/constants/place-offers.ts';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
